server/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
sqlalchemy>=2.0.40
passlib>=1.7.4
PyJWT>=2.10.1
python-multipart>=0.0.20
